@@ -1,155 +1,177 @@
 {
-	"name": "play-zero",
-	"description": "a dao for videogames — landingpage",
-	"version": "0.1.2",
-	"license": "UNLICENSED",
-	"scripts": {
-		"dev": "next -p 3030",
-		"build": "next build",
-		"start": "next start",
-		"export": "next export",
-		"format": "prettier --loglevel error --write 'src/**/*.{js,json,jsx,ts,tsx}'",
-		"format:verify": "prettier --loglevel error --list-different 'src/**/*.{js,json,jsx,ts,tsx}'",
-		"test": "echo test",
-		"codegen": "graphql-codegen --config codegen.yml",
-		"patch": "yarn version --patch --no-git-tag-version"
-	},
-	"dependencies": {
-		"3box": "^1.19.0",
-		"@apollo/react-hooks": "^3.1.5",
-		"@mdx-js/loader": "^1.6.4",
-		"@next/mdx": "^9.4.2",
-		"@polkadot/api": "^1.14.1",
-		"@polkadot/extension-dapp": "^0.25.1",
-		"@polkadot/keyring": "^2.10.1",
-		"@polkadot/ui-keyring": "^0.53.1",
-		"@polkadot/util": "^2.10.1",
-		"@polkadot/util-crypto": "^2.10.1",
-		"@rebass/forms": "^4.0.6",
-		"@rebass/preset": "^4.0.5",
-		"@sendgrid/mail": "^7.1.1",
-		"apollo-cache": "^1.3.5",
-		"apollo-cache-inmemory": "^1.6.6",
-		"apollo-client": "^2.6.10",
-		"apollo-link": "^1.2.14",
-		"apollo-link-context": "^1.0.20",
-		"apollo-link-error": "^1.1.13",
-		"apollo-link-http": "^1.5.17",
-		"apollo-link-ws": "^1.0.20",
-		"apollo-server-micro": "^2.13.1",
-		"apollo-utilities": "^1.3.4",
-		"body-parser": "^1.19.0",
-		"cors": "^2.8.5",
-		"eth-sig-util": "^2.5.3",
-		"ethereumjs-util": "^7.0.1",
-		"express": "^4.17.1",
-		"express-jwt": "^5.3.3",
-		"fathom-client": "^3.0.0",
-		"formik": "^2.1.4",
-		"graphql": "^15.0.0",
-		"graphql-type-json": "^0.3.1",
-		"isomorphic-unfetch": "^3.0.0",
-		"micro": "^9.3.4",
-		"micro-cors": "^0.1.1",
-		"mixpanel-browser": "^2.36.0",
-		"next": "^9.4.2",
-		"next-images": "^1.4.0",
-		"next-page-transitions": "^1.0.0-beta.2",
-		"next-with-apollo": "^5.0.1",
-		"nextjs-sitemap-generator": "^1.0.0",
-		"react": "^16.13.1",
-		"react-apollo": "^3.1.5",
-		"react-dom": "^16.13.1",
-		"react-icons": "^3.10.0",
-		"react-mailchimp-subscribe": "^2.1.0",
-		"react-tracking": "^7.3.0",
-		"react-ui": "^1.0.0-beta.26",
-		"rebass": "^4.0.7",
-		"sequelize": "^5.21.10",
-		"styled-components": "^5.1.0",
-		"subscriptions-transport-ws": "^0.9.16"
-	},
-	"devDependencies": {
-		"@commitlint/cli": "^8.3.5",
-		"@commitlint/config-conventional": "^8.3.4",
-		"@commitlint/prompt-cli": "^8.3.5",
-		"@types/node": "^14.0.4",
-		"@types/react": "16.9.35",
-		"@types/react-dom": "16.9.8",
-		"@types/react-mailchimp-subscribe": "^2.0.2",
-		"@types/react-tracking": "^7.0.2",
-		"@types/rebass": "^4.0.6",
-		"@types/styled-components": "5.1.0",
-		"babel-plugin-module-resolver": "^4.0.0",
-		"babel-plugin-styled-components": "^1.10.7",
-		"conventional-changelog": "^3.1.21",
-		"dotenv": "^8.2.0",
-		"husky": "^4.2.5",
-		"jest": "^26.0.1",
-		"lint-staged": "^10.2.4",
-		"prettier": "^2.0.5",
-		"pretty-quick": "^2.0.1",
-		"ts-jest": "^26.0.0",
-		"typescript": "^3.9.3",
-		"url-loader": "^4.1.0"
-	},
-	"jest": {
-		"moduleFileExtensions": [
-			"ts",
-			"tsx",
-			"js"
-		],
-		"transform": {
-			"^.+\\.tsx?$": "ts-jest"
-		},
-		"testMatch": [
-			"**/*.(test|spec).(ts|tsx)"
-		],
-		"globals": {
-			"ts-jest": {
-				"babelConfig": true,
-				"tsConfig": "jest.tsconfig.json"
-			}
-		},
-		"coveragePathIgnorePatterns": [
-			"/node_modules/",
-			"enzyme.js"
-		],
-		"setupTestFrameworkScriptFile": "<rootDir>/enzyme.js",
-		"coverageReporters": [
-			"json",
-			"lcov",
-			"text",
-			"text-summary"
-		],
-		"moduleNameMapper": {
-			"\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/mocks.js",
-			"\\.(css|less)$": "<rootDir>/__mocks__/mocks.js"
-		}
-	},
-	"husky": {
-		"hooks": {
-			"commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-			"pre-commit": "lint-staged"
-		}
-	},
-	"prettier": {
-		"trailingComma": "es5",
-		"tabWidth": 4,
-		"useTabs": true,
-		"semi": false,
-		"singleQuote": true,
-		"printWidth": 160,
-		"bracketSpacing": true
-	},
-	"lint-staged": {
-		"*.{js,ts,jsx,tsx,json,css}": [
-			"yarn format"
-		]
-	},
-	"commitlint": {
-		"extends": [
-			"@commitlint/config-conventional"
-		]
-	}
+  "name": "gamedao",
+  "description": "community owned crowdfinance cooperative dao for videogames, content and esports",
+  "version": "0.1.0",
+  "author": [
+    "2075@zero.io"
+  ],
+  "license": "Unlicense",
+  "private": true,
+  "scripts": {
+    "postinstall": "yarn patch",
+    "dev": "next -p 3333",
+    "build": "next build",
+    "patch+build": "yarn patch && next build",
+    "start": "next start",
+    "export": "next export",
+    "format": "prettier --write --check 'src/**/*.{js,json,jsx,ts,tsx}'",
+    "format:verify": "prettier --check 'src/**/*.{js,json,jsx,ts,tsx}'",
+    "test": "echo test",
+    "codegen": "graphql-codegen --config codegen.yml",
+    "patch": "./patch/patch.sh",
+    "build:types": "yarn fetch:meta && yarn generate:defs && yarn generate:meta",
+    "fetch:meta": "curl -H \"Content-Type: application/json\" -d '{\"id\":\"1\", \"jsonrpc\":\"2.0\", \"method\": \"state_getMetadata\", \"params\":[]}' http://alphaville.zero.io:9933 > zero-metadata.json",
+    "generate:defs": "ts-node --skip-project node_modules/.bin/polkadot-types-from-defs --package sample-polkadotjs-typegen/interfaces --input ./src/interfaces",
+    "generate:meta": "ts-node --skip-project node_modules/.bin/polkadot-types-from-chain --package sample-polkadotjs-typegen/interfaces --endpoint ./zero-metadata.json --output ./src/interfaces"
+  },
+  "peerDependencies": {},
+  "dependencies": {
+    "@apollo/link-context": "^2.0.0-beta.3",
+    "@apollo/link-ws": "^2.0.0-beta.3",
+    "@apollo/react-hooks": "^4.0.0",
+    "@mdx-js/loader": "^1.6.21",
+    "@next/mdx": "^10.1.3",
+    "@polkadot/api": "^4.9.2",
+    "@polkadot/extension-dapp": "^0.37.2",
+    "@polkadot/keyring": "^6.4.1",
+    "@polkadot/typegen": "^4.9.2",
+    "@polkadot/types": "^4.9.2",
+    "@polkadot/ui-keyring": "^0.76.1",
+    "@polkadot/ui-settings": "^0.76.1",
+    "@polkadot/util": "^6.4.1",
+    "@polkadot/util-crypto": "^6.4.1",
+    "@rebass/forms": "^4.0.6",
+    "@rebass/preset": "^4.0.5",
+    "apollo-client": "^2.6.10",
+    "apollo-link-context": "^1.0.20",
+    "apollo-server-micro": "^2.22.2",
+    "body-parser": "^1.19.0",
+    "chalk": "^4.1.0",
+    "cors": "^2.8.5",
+    "discord.js": "^12.5.3",
+    "express": "^4.17.1",
+    "express-jwt": "^6.0.0",
+    "fathom-client": "^3.0.0",
+    "formik": "^2.2.5",
+    "graphql": "^15.5.0",
+    "graphql-type-json": "^0.3.1",
+    "isomorphic-unfetch": "^3.0.0",
+    "lerna": "^4.0.0",
+    "localforage": "^1.9.0",
+    "loglevel": "^1.7.1",
+    "loglevel-plugin-prefix": "^0.8.4",
+    "micro": "^9.3.4",
+    "micro-cors": "^0.1.1",
+    "next": "^10.1.3",
+    "next-images": "^1.7.0",
+    "next-page-transitions": "^1.0.0-beta.2",
+    "nextjs-sitemap-generator": "^1.3.1",
+    "nodemailer": "^6.5.0",
+    "query-string": "^7.0.0",
+    "react": "^17.0.2",
+    "react-copy-to-clipboard": "^5.0.3",
+    "react-dom": "^17.0.2",
+    "react-icons": "^4.2.0",
+    "react-mailchimp-subscribe": "^2.1.3",
+    "rebass": "^4.0.7",
+    "semantic-ui-css": "2.4.1",
+    "semantic-ui-react": "2.0.3",
+    "sequelize": "^6.6.2",
+    "styled-components": "^5.2.3",
+    "toformat": "^2.0.0",
+    "typescript": "^4.2.3"
+  },
+  "devDependencies": {
+    "@commitlint/cli": "^12.1.1",
+    "@commitlint/config-conventional": "^12.1.1",
+    "@commitlint/prompt-cli": "^12.1.1",
+    "@jest/core": "^26.6.3",
+    "@picgo/bump-version": "^1.0.3",
+    "@types/jest": "^26.0.22",
+    "@types/node": "14.14.37",
+    "@types/react": "^17.0.3",
+    "@types/react-dom": "^17.0.3",
+    "@types/rebass": "^4.0.8",
+    "@types/styled-components": "5.1.9",
+    "babel-plugin-styled-components": "^1.10.7",
+    "commitizen": "^4.2.3",
+    "concurrently": "^6.0.0",
+    "conventional-changelog": "^3.1.24",
+    "cz-customizable": "^6.2.0",
+    "dotenv": "^8.2.0",
+    "husky": "^6.0.0",
+    "jest": "^26.6.3",
+    "jest-cli": "^26.6.3",
+    "lint-staged": "^10.5.4",
+    "prettier": "^2.2.1",
+    "pretty-quick": "^3.1.0",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.5.4",
+    "ts-node": "^9.1.1",
+    "url-loader": "^4.0.0",
+    "webpack": "^5.30.0"
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-customizable"
+    },
+    "cz-customizable": {
+      "config": "./node_modules/@picgo/bump-version/.cz-config.js"
+    }
+  },
+  "jest": {
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx"
+    ],
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testMatch": [
+      "**/*.(test|spec).(ts|tsx)"
+    ],
+    "globals": {
+      "ts-jest": {
+        "babelConfig": true,
+        "tsConfig": "jest.tsconfig.json"
+      }
+    },
+    "coveragePathIgnorePatterns": [
+      "/node_modules/",
+      "enzyme.js"
+    ],
+    "setupTestFrameworkScriptFile": "<rootDir>/enzyme.js",
+    "coverageReporters": [
+      "json",
+      "lcov",
+      "text",
+      "text-summary"
+    ],
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/mocks.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/mocks.js"
+    }
+  },
+  "husky": {
+    "hooks": {
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "yarn format:verify"
+    }
+  },
+  "commitlint": {
+    "extends": [
+      "@commitlint/config-conventional"
+    ]
+  },
+  "prettier": {
+    "trailingComma": "es5",
+    "tabWidth": 4,
+    "useTabs": true,
+    "semi": false,
+    "singleQuote": true,
+    "printWidth": 160,
+    "bracketSpacing": true
+  }
 }
