@@ -9,23 +9,41 @@
            © C O P Y R I O T   2 0 7 5   Z E R O . I O
 **/
 
-const sgMail = require('@sendgrid/mail')
+// const sgMail = require('@sendgrid/mail')
+const nodemailer = require('nodemailer')
+import Discord from 'discord.js'
+
+const ID = '748193288997044395'
+const TOKEN = 'U5pvMChmhau5zLb8GHUs7MKHUQ_0qIMfVAnMtv0Qgt4n6pOdMNrUNSUKq6c7S-ZpAsFX'
+const discord = new Discord.WebhookClient(ID, TOKEN)
 
 export default async function (req, res) {
-	sgMail.setApiKey(process.env.SENDGRID_API_KEY || '')
+	// sgMail.setApiKey(process.env.SENDGRID_API_KEY || '')
+
+	let transporter = nodemailer.createTransport({
+		host: 'mail.zero.io',
+		port: 465,
+		secure: true,
+		auth: {
+			user: 'play@zero.io',
+			pass: 'frgGDgHBme6mmVM',
+		},
+	})
 
 	const { name, email, phone, message } = req.body
 
 	const content = {
-		to: 'play+contact@zero.io',
+		to: 'play@zero.io',
 		from: email,
-		subject: `ZERO.IO — Message from ${name || email}`,
+		subject: `GameDAO — Message from ${name || email}`,
 		text: `from:\nname: ${name}\n email: ${email}\nphone: ${phone}\n\n${message}`,
 		html: `<p>from:</p><p>name: ${name}</p><p>email: ${email}</p><p>phone: ${phone}</p><p>${message}</p>`,
 	}
 
 	try {
-		await sgMail.send(content)
+		discord.send('```' + content.text + '```')
+		await transporter.sendMail(content)
+		// await sgMail.send(content)
 		res.status(200).send('Thank you. The message was successfully sent.')
 	} catch (error) {
 		console.log('ERROR', error)
