@@ -1,7 +1,7 @@
-import { FunctionComponent } from 'react'
+// import { FunctionComponent } from 'react'
 
-declare module 'react' {
-	export interface NextFunctionComponent<P> extends FunctionComponent<P> {
-		getInitialProps(ctx): Promise<P>
-	}
-}
+// declare module 'react' {
+// 	export interface NextFunctionComponent<P> extends FunctionComponent<P> {
+// 		getInitialProps(ctx): Promise<P>
+// 	}
+// }
