@@ -1,8 +1,14 @@
 import React, { useState, useEffect } from 'react'
 import styled from 'styled-components'
 
-import { Flex, Box, Heading, Text } from 'rebass'
-import { H1, Link, Intro, Container, ContentGrid } from 'components'
+import { Flex, Box, Heading, Text } from 'rebass/styled-components'
+import {
+	H1,
+	Link,
+	Intro,
+	Container,
+	// ContentGrid
+} from 'components'
 
 import img from 'public/img/1.png'
 import logo from 'public/z-ctrl-45-wht.svg'
@@ -19,7 +25,7 @@ const Page = () => {
 			</Flex>
 			<Flex id="content" flexDirection="row" flexWrap="wrap" color="rgba(255,255,255,.65)">
 				<Box width={[1]} px={[2, 4]} py={[2, 4]}>
-					<ContentGrid />
+					{/*<ContentGrid />*/}
 				</Box>
 			</Flex>
 		</Container>
