@@ -1,3 +1,14 @@
+/**
+           _______________________________ ________
+           \____    /\_   _____/\______   \\_____  \
+             /     /  |    __)_  |       _/ /   |   \
+            /     /_  |        \ |    |   \/    |    \
+           /_______ \/_______  / |____|_  /\_______  /
+                   \/        \/         \/         \/
+           Z  E  R  O  .  I  O     N  E  T  W  O  R  K
+           © C O P Y R I O T   2 0 7 5   Z E R O . I O
+**/
+
 import React, { useContext } from 'react'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
@@ -6,7 +17,7 @@ import styled from 'styled-components'
 import { Flex, Box } from 'rebass'
 import { Header, Footer, Content, H1 } from 'components'
 
-import { AppContext } from 'src/hooks/AppContext'
+import { AppContext } from 'src/context/AppContext'
 
 const Wrapper = styled.div`
 	margin: 0;
@@ -28,7 +39,9 @@ const Layout = (props) => {
 	const { state } = useContext(AppContext)
 	const { SITE_NAME, SITE_TITLE, SITE_DESCRIPTION, SITE_IMAGE, TW_SITE_CREATOR, TW_SITE_NAME } = state.config.data
 	const { SHOW_HEADER, SHOW_FOOTER } = state.features.data
+
 	const path = router.pathname.substr(1, 1).toUpperCase() + router.pathname.substr(2)
+	const showDiscord = path === ''
 
 	return (
 		<Wrapper>
@@ -55,6 +68,22 @@ const Layout = (props) => {
 				<title>
 					{path || 'GameDAO'} | {SITE_TITLE}
 				</title>
+
+				{showDiscord && (
+					<script
+						src="https://cdn.jsdelivr.net/npm/@widgetbot/crate@3"
+						async
+						defer
+						dangerouslySetInnerHTML={{
+							__html: `
+								new Crate({
+									server: '273529551483699200', // ZΞRO
+									channel: '732569366431924248' // #🪐web
+								})
+							`,
+						}}
+					/>
+				)}
 			</Head>
 			<Viewport>
 				{SHOW_HEADER && <Header />}
