@@ -1,17 +1,12 @@
-
 export const shuffle = <T>(array: T[]) => {
-
 	const shuffled = [...array]
 
-	for ( let i = shuffled.length - 1; i > 0; i--) {
-
+	for (let i = shuffled.length - 1; i > 0; i--) {
 		const j = Math.floor(Math.random() * i)
 		const temp = shuffled[i]
 		shuffled[i] = shuffled[j]
 		shuffled[j] = temp
-
 	}
 
 	return shuffled
-
 }
