@@ -14,10 +14,12 @@ import * as React from 'react'
 import Layout from 'layouts'
 import { Home } from 'containers'
 
-const Page = () => (
-	<Layout>
-		<Home />
-	</Layout>
-)
+const Page = (props) => {
+	return (
+		<Layout>
+			<Home />
+		</Layout>
+	)
+}
 
 export default Page
